Split to separate classes and traits (#83)
