+++ ingestor-default/src/main/scala/za/co/absa/hyperdrive/ingestor/implementation/decoder/avro/confluent/ConfluentAvroKafkaStreamDecoder.scala
@@ -21,10 +21,9 @@ import org.apache.logging.log4j.LogManager
 import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.streaming.DataStreamReader
 import za.co.absa.abris.avro.read.confluent.SchemaManager
-import za.co.absa.abris.avro.read.confluent.SchemaManager.{PARAM_SCHEMA_NAMESPACE_FOR_RECORD_STRATEGY, PARAM_SCHEMA_NAME_FOR_RECORD_STRATEGY, PARAM_SCHEMA_REGISTRY_URL, PARAM_VALUE_SCHEMA_ID, PARAM_VALUE_SCHEMA_NAMING_STRATEGY, SchemaStorageNamingStrategies}
-import za.co.absa.hyperdrive.ingestor.api.PropertyMetadata
-import za.co.absa.hyperdrive.ingestor.api.decoder.{StreamDecoder, StreamDecoderFactory, StreamDecoderFactoryProvider}
-import za.co.absa.hyperdrive.shared.configurations.ConfigurationsKeys.AvroKafkaStreamDecoderKeys.{KEY_SCHEMA_REGISTRY_URL, KEY_SCHEMA_REGISTRY_VALUE_NAMING_STRATEGY, KEY_SCHEMA_REGISTRY_VALUE_RECORD_NAME, KEY_SCHEMA_REGISTRY_VALUE_RECORD_NAMESPACE, KEY_SCHEMA_REGISTRY_VALUE_SCHEMA_ID, KEY_TOPIC}
+import za.co.absa.abris.avro.read.confluent.SchemaManager.{PARAM_SCHEMA_NAMESPACE_FOR_RECORD_STRATEGY, PARAM_SCHEMA_NAME_FOR_RECORD_STRATEGY, PARAM_VALUE_SCHEMA_NAMING_STRATEGY, SchemaStorageNamingStrategies}
+import za.co.absa.hyperdrive.ingestor.api.decoder.{StreamDecoder, StreamDecoderFactory}
+import za.co.absa.hyperdrive.shared.configurations.ConfigurationsKeys.AvroKafkaStreamDecoderKeys._
 import za.co.absa.hyperdrive.shared.utils.ConfigUtils.getOrThrow
 
 private[decoder] class ConfluentAvroKafkaStreamDecoder(val topic: String, val schemaRegistrySettings: Map[String,String]) extends StreamDecoder {
@@ -51,8 +50,8 @@ private[decoder] class ConfluentAvroKafkaStreamDecoder(val topic: String, val sc
     val schemaRegistryFullSettings = schemaRegistrySettings + (SchemaManager.PARAM_SCHEMA_REGISTRY_TOPIC -> topic)
     logger.info(s"SchemaRegistry settings: $schemaRegistryFullSettings")
 
-    import za.co.absa.abris.avro.functions.from_confluent_avro
     import org.apache.spark.sql.functions.col
+    import za.co.absa.abris.avro.functions.from_confluent_avro
     streamReader
       .load()
       .select(from_confluent_avro(col("value"), schemaRegistryFullSettings) as 'data)
@@ -60,7 +59,7 @@ private[decoder] class ConfluentAvroKafkaStreamDecoder(val topic: String, val sc
   }
 }
 
-object ConfluentAvroKafkaStreamDecoder extends StreamDecoderFactory {
+object ConfluentAvroKafkaStreamDecoder extends StreamDecoderFactory with ConfluentAvroKafkaStreamDecoderAttributes {
 
   override def apply(config: Configuration): StreamDecoder = {
     val topic = getTopic(config)
@@ -70,19 +69,6 @@ object ConfluentAvroKafkaStreamDecoder extends StreamDecoderFactory {
 
     new ConfluentAvroKafkaStreamDecoder(topic, schemaRegistrySettings)
   }
-
-  override def getName: String = "Confluent Avro Stream Decoder"
-
-  override def getDescription: String = "Decoder for Kafka messages in Avro format. The decoder connects to a Schema Registry instance to retrieve the schema information."
-
-  override def getProperties: Map[String, PropertyMetadata] = Map(
-    PARAM_SCHEMA_REGISTRY_URL -> PropertyMetadata("Schema Registry URL", None, required = true),
-    PARAM_VALUE_SCHEMA_ID -> PropertyMetadata("Schema Id", Some("Specific Id of schema or \"latest\""), required = true),
-    PARAM_VALUE_SCHEMA_NAMING_STRATEGY -> PropertyMetadata("Schema naming strategy",
-      Some("Subject name strategy of Schema Registry. Must be one of \"topic.name\", \"record.name\" or \"topic.record.name\""), required = true),
-    PARAM_SCHEMA_NAME_FOR_RECORD_STRATEGY -> PropertyMetadata("Record name", Some("Record name for naming strategies record.name or topic.record.name"), required = false),
-    PARAM_SCHEMA_NAMESPACE_FOR_RECORD_STRATEGY -> PropertyMetadata("Record namespace", Some("Record namespace for naming strategies record.name or topic.record.name"), required = false)
-  )
 
   private def getTopic(configuration: Configuration): String = getOrThrow(KEY_TOPIC, configuration, errorMessage = s"Topic not found. Is '$KEY_TOPIC' properly set?")
 
@@ -114,8 +100,4 @@ object ConfluentAvroKafkaStreamDecoder extends StreamDecoderFactory {
     import SchemaStorageNamingStrategies._
     strategy == RECORD_NAME || strategy == TOPIC_RECORD_NAME
   }
-}
-
-class ConfluentAvroKafkaStreamDecoderLoader extends StreamDecoderFactoryProvider {
-  override def getComponentFactory: StreamDecoderFactory = ConfluentAvroKafkaStreamDecoder
 }
