+++ ingestor-default/src/main/scala/za/co/absa/hyperdrive/ingestor/implementation/reader/kafka/KafkaStreamReader.scala
@@ -82,7 +82,7 @@ private[reader] class KafkaStreamReader(val topic: String, val brokers: String, 
   }
 }
 
-object KafkaStreamReader extends StreamReaderFactory {
+object KafkaStreamReader extends StreamReaderFactory with KafkaStreamReaderAttributes {
   private val logger = LogManager.getLogger
 
   override def apply(conf: Configuration): StreamReader = {
@@ -94,16 +94,6 @@ object KafkaStreamReader extends StreamReaderFactory {
 
     new KafkaStreamReader(topic, brokers, extraOptions)
   }
-
-  override def getName: String = "Kafka Reader"
-
-  override def getDescription: String = "This component ingests data from a kafka topic."
-
-  override def getProperties: Map[String, PropertyMetadata] = Map(
-    KEY_TOPIC -> PropertyMetadata("Topic name", Some("Name of the kafka topic"), required = true),
-    KEY_BROKERS -> PropertyMetadata("Brokers", Some("Comma-separated list of kafka broker urls"), required = true))
-
-  override def getExtraConfigurationPrefix: Option[String] = Some(rootFactoryOptionalConfKey)
 
   private def getTopic(configuration: Configuration): String = {
     getOrThrow(KEY_TOPIC, configuration, errorMessage = s"Topic not found. Is $KEY_TOPIC defined?")
@@ -117,8 +107,4 @@ object KafkaStreamReader extends StreamReaderFactory {
   private def getExtraOptions(configuration: Configuration): Map[String, String] = ConfigUtils.getPropertySubset(configuration, rootFactoryOptionalConfKey)
 
   private def filterKeysContaining(map: Map[String, String], exclusionToken: String) = map.filterKeys(!_.contains(exclusionToken))
-}
-
-class KafkaStreamReaderLoader extends StreamReaderFactoryProvider {
-  override def getComponentFactory: StreamReaderFactory = KafkaStreamReader
 }
