+++ ingestor-default/src/main/scala/za/co/absa/hyperdrive/ingestor/implementation/reader/kafka/KafkaStreamReaderAttributes.scala
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2018 ABSA Group Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package za.co.absa.hyperdrive.ingestor.implementation.reader.kafka
+
+import za.co.absa.hyperdrive.ingestor.api.{HasComponentAttributes, PropertyMetadata}
+import za.co.absa.hyperdrive.shared.configurations.ConfigurationsKeys.KafkaStreamReaderKeys.{KEY_BROKERS, KEY_TOPIC, rootFactoryOptionalConfKey}
+
+trait KafkaStreamReaderAttributes extends HasComponentAttributes {
+
+  override def getName: String = "Kafka Reader"
+
+  override def getDescription: String = "This component ingests data from a kafka topic."
+
+  override def getProperties: Map[String, PropertyMetadata] = Map(
+    KEY_TOPIC -> PropertyMetadata("Topic name", Some("Name of the kafka topic"), required = true),
+    KEY_BROKERS -> PropertyMetadata("Brokers", Some("Comma-separated list of kafka broker urls"), required = true))
+
+  override def getExtraConfigurationPrefix: Option[String] = Some(rootFactoryOptionalConfKey)
+}
